XAML Bringup - skeleton DdiResourceCopyRegion11_1
This changelist includes a skeleton implementation, including parameter checking for ResourceCopyRegion11_1 per MSDN.

TODOs are left in the code for parts of the implementation and paramater checking that have not been hit in the driver, yet.

This changelist also shuffles a bit of code in SetPredication based on some learnings after plumbing ResourceCopyRegion.